Renamed files to Jira bug number
