+++ tests/Doctrine/Tests/Models/DDC2376/Reference.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Doctrine\Tests\Models\AddManyToOne;
+namespace Doctrine\Tests\Models\DDC2376;
 
 use Doctrine\ORM\Mapping\ClassMetadata;
 use Doctrine\ORM\Mapping\Builder\ClassMetadataBuilder;
@@ -13,7 +13,7 @@ class Reference
         $builder->createField('id', 'integer')->isPrimaryKey()->generatedValue()->build();
         $builder->addField('name', 'string');
 
-        $builder->addManyToOne('user', 'Doctrine\Tests\Models\AddManyToOne\User');
+        $builder->addManyToOne('user', 'Doctrine\Tests\Models\DDC2376\User');
     }
 
     protected $id;
